Rename files to TextInput
