+++ apps/test-app/app/routes/tests/text-input/index.spec.ts-snapshot.-visual-1-webkit-linux.png
