+++ apps/test-app/app/routes/tests/text-input/index.tsx
@@ -6,7 +6,7 @@ import { TextInput, Label } from "@itwin/kiwi-react/bricks";
 import { useSearchParams } from "@remix-run/react";
 import { useId } from "react";
 
-export const handle = { title: "Input" };
+export const handle = { title: "TextInput" };
 
 export default function Page() {
 	const disabled = useSearchParams()[0].get("disabled") === "true";
