+++ packages/kiwi-react/src/bricks/index.ts
@@ -10,12 +10,12 @@ export * as DropdownMenu from "./DropdownMenu.js";
 export { Divider } from "./Divider.js";
 export { Icon } from "./Icon.js";
 export { IconButton } from "./IconButton.js";
-export { TextInput } from "./Input.js";
 export { Field } from "./Field.js";
 export { Label } from "./Label.js";
 export { Radio } from "./Radio.js";
 export { Switch } from "./Switch.js";
 export * as Tabs from "./Tabs.js";
 export { Textarea } from "./Textarea.js";
+export { TextInput } from "./TextInput.js";
 export { Tooltip } from "./Tooltip.js";
 export { VisuallyHidden } from "./VisuallyHidden.js";
