+++ packages/kiwi-react/src/bricks/styles.css
@@ -12,11 +12,11 @@
 @import "./Divider.css" layer(kiwi.components);
 @import "./DropdownMenu.css" layer(kiwi.components);
 @import "./IconButton.css" layer(kiwi.components);
-@import "./Input.css" layer(kiwi.components);
 @import "./Label.css" layer(kiwi.components);
 @import "./Radio.css" layer(kiwi.components);
 @import "./ListItem.css" layer(kiwi.components);
 @import "./Switch.css" layer(kiwi.components);
+@import "./TextInput.css" layer(kiwi.components);
 @import "./Tooltip.css" layer(kiwi.components);
 
 @import "./Field.css" layer(kiwi.components);
